Refine cheated endpoint_to_payload semantics reflection proof
This uses the local confluence property of the payload semantics
to reduce the goal so that we only need to mimic target-language
steps starting from states that are in the range of the compilation
function, rather than steps starting from (arbitrarily interleaved)
intermedate steps.

This, I hope, completely eliminates the need for the painful
invariant that my previous sketch of this proof used.