Make payload semantics maintain normalised queues
James' queue representation entails that qcong-related queues
have the same representation, with one exception: the empty
queue does not have a unique representation.

An empty queue can be represented as FEMPTY, as
FEMPTY |+ (p,[]), and so on.

This is an experiment with changing the payload semantics
so that it maintains the invariant that queues are always
normalised, in the sense that empty queues are represented
as FEMPTY.