Alternate version of compilation_preservation
This is not as big of an imporvement as I hoped. I was hoping
it would be possible to not mention the BISIM_REL on the
payloadLang state at all. I still believe this is possible,
but it would be a daunting amount of work:
one would have to prove that CakeML evaluation of compiled
endpoints that are related by permutation and/or junkcong
agrees, and (I think) that bisimilarity is preserve by
net_filter.