Update strings to variable names (issue #14)
There's still some work to do in the payload_to_cakeml proofs,
but this commit updates everything else.

Almost all broken proofs were due to generated names changing
in syntax inductions. In a few cases, I've tried to remove
the generated name dependencies, but in many I've just
alpha-converted the proof and let it be.