Sketch of fixpoints->closures compiler phase
A complication here is that closures can only be invoked with defined
variables, but fixpoints don't care whether variables in its body are
defined or not. For example, consider something like the following:

  let x = 0 in
    fix X. if x = 0 then y:= 1; x:=1; X
           else Send y to p

To convert this fixpoint to a closure, it would be necessary to pass
both x and y as parameters because these variables change during the
function body.  However, the first time this fixpoint is unfolded, y
is undefined, and we have no value to pass to the closure

To work around this, the compilation makes sure every variable that
can potentially be written to is initialised.

The procedure that selects which arguments should be passed to a closure
do not account for writes incurred as a result of nested fixpoints in
any special way. It's not clear to me whether anything special needs
to be done to account for, e.g., programs like the following:

  fix X.
    v := v+1;
    fix Y.
      if ... then
        v' := v'+1; Y
      else
        X