Clean up previous error-freedom proof attempt
The previous approach cannot actually work because it made the
needlessly strong assumption that there are no errors whatsoever in
the iforest; that's not true of chor forests. They have errors,
they're just hopefully unreachable.

For now, there's a proof that choreographies do not project to error
states.  Which doesn't mean they'll never reach them, just that they
can eventually be rid of them.